+++ PHP/view-projecten.php
@@ -41,43 +41,64 @@ session_start();
                 <div class="sidebar-brand-text mx-3">AFA</div>
             </a>
             <!-- Divider -->
-            <hr class="sidebar-divider my-0">
+                <hr class="sidebar-divider my-0">
             <!-- Nav Item - Dashboard -->
-            <li class="nav-item ">
-                <a class="nav-link" href="../home.php">
-                    <i class="fas fa-tachometer-alt"></i>
-                    <span>Dashboard</span></a>
-            </li>
-            <hr class="sidebar-divider my-0">
-            <li class="nav-item active">
-                <a class="nav-link" href="../administratie.php">
+                <li class="nav-item ">
+                    <a class="nav-link" href="../home.php">
+                        <i class="fas fa-tachometer-alt"></i>
+                        <span>Dashboard</span>
+                    </a>
+                </li>
+            <!-- Divider -->
+                <hr class="sidebar-divider my-0">
+            <!--Nav Item - Projecten-->        
+                <li class="nav-item active">
+                    <a class="nav-link" href="../administratie.php">
                     <i class="fas fa-file-medical"></i>
                     <?php
-      
-        if ($_SESSION['role'] == 'Administratie'or $_SESSION['role'] == 'Beheerder'){
-           ?>
-
-                    <span>Registreer Projecten</span></a>
-
+                        if ($_SESSION['role'] == 'Administratie' or $_SESSION['role'] == 'Beheerder') 
+                        {
+                    ?>
+                        <span>Registreer Projecten</span></a>
+                    <?php
+                        }
+                        else
+                        {
+                    ?>
+                            <span>Overzicht Projecten</span></a>
+                    <?php
+                        }
+                    ?>
+                </li>
+            <!--Nav Item - Personen-->
                 <?php
-        }else{
-        ?>
-                <span>Projecten</span></a>
-                <?php } ?>
-            </li>
-            <?php
-      
-        if ($_SESSION['role'] == 'Administratie'or $_SESSION['role'] == 'Beheerder'){ ?>
-            <li class="nav-item">
-                <a class="nav-link" href="../administratie-personen.php">
-                    <i class="fas fa-user-plus"></i>
-                    <span>Registreer Personen</span></a>
-            </li>
-            <?php
-    }
-    ?>
-            </li>
-
+                    if ($_SESSION['role'] == 'Administratie' or $_SESSION['role'] == 'Beheerder') 
+                    {
+                ?>
+                        <li class="nav-item">
+                            <a class="nav-link" href="../administratie-personen.php">
+                                <i class="fas fa-user-edit"></i>
+                                <span>Registreer Personen</span>
+                            </a>
+                        </li>
+                <?php
+                    }
+                ?>
+            <!--Nav Item - Gebruikers-->
+                <?php
+                    if ($_SESSION['role'] == 'Beheerder') 
+                    {
+                ?>
+                        <li class="nav-item">
+                            <a class="nav-link" href="../Gebruikers.php">
+                                <i class="fas fa-user-edit"></i>
+                                <span>Registreer Gebruikers</span>
+                            </a>
+                        </li>
+                <?php
+                    }
+                ?>
+        <!--Nav Items - END-->
             <!-- Sidebar Toggler (Sidebar) -->
             <div class="text-center d-none d-md-inline">
                 <button class="rounded-circle border-0" id="sidebarToggle"></button>
@@ -264,11 +285,20 @@ left join taak on  bestedingen.TaakID = taak.ID
     <div class='additional'>
       <div class='user-card'>";
       // if($status=="Niet Compleet"){
-        echo "<a class='link' href='registratie-bestedingen.php?id=$id&idt=$idt'><button class='icon5'  data-role='update' data-id='$idt'><i class='fas fa-eye' data-toggle='tooltip' data-placement='top' title='View'></i></button></a>
-              <a class='link' href='#'><button class='icon' onclick=EditTaak($idt)><i class='fas fa-edit' data-toggle='tooltip' data-placement='top' title='Edit'></i></button></a>
-              <a class='link' href='#'><button class='icon6' onclick=DeleteTaak($idt)><i class='fas fa-trash-alt' data-toggle='tooltip' data-placement='top' title='Delete'></i></button></a>
-               
-        ";
+        echo "
+                <a class='link' href='registratie-bestedingen.php?id=$id&idt=$idt'><button class='icon5'  data-role='update' data-id='$idt'><i class='fas fa-eye' data-toggle='tooltip' data-placement='top' title='View'></i></button></a>
+             ";  
+                
+            if ($_SESSION['role'] == 'Administratie' or $_SESSION['role'] == 'Beheerder'){
+                echo "
+                        <a class='link' href='#'><button class='icon' onclick=EditTaak($idt)><i class='fas fa-edit' data-toggle='tooltip' data-placement='top' title='Edit'></i></button></a>
+                    ";
+            }
+            if ($_SESSION['role'] == 'Beheerder'){
+                echo "
+                        <a class='link' href='#'><button class='icon6' onclick=DeleteTaak($idt)><i class='fas fa-trash-alt' data-toggle='tooltip' data-placement='top' title='Delete'></i></button></a>
+                     ";
+            }
 
       // }else{
